feat: Introduce l10n
There are some icons which are specifically for a country, like Belgium
or The Netherlands. These are mixed within the icon sets.

There has been a request to separate them from the rest of the icons,
and put them in a separate folder, specifically for that country. This
way a list can be made from the icons within one folder and overlay
the icons with a country icon if needed.

This patch introduces localization (l10n). The country specific icons
are moved into a dedicated folder for each collection using the
ISO_3166-1 (alpha-2) standard for the folder names. In each collection
the icons `PaspoortIdkaartGecombineerd<country_name>.svg` and
`Paspoort<country_name>.svg` have moved to these dedicated folders and
renamed to their regular (international) variant.