feat: Introduce t9n
Currently the title within the svg’s is in Dutch. Also the names of the
icons are in Dutch, as uppercamelcase. You can use these to generate a
list of all the icons, for instance to make a select box in a CMS to
select an icon.

There has been a request to provide XLF files for the translation of
these names.

This patch provides the XLF filee starting with the default English
(still must be translated from Dutch to English), Dutch and German